Fixed bug in type mismatching(savannah bug 64670). Default data type set to 2009
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@22185 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 6b5f4af557eb3276b46e798f4a43b94c5350ce35

+++ Rec/Muon/MuonID/src/MuonIDAlg.h
@@ -138,7 +138,7 @@ private:
   
   // GL&SF: 
   bool m_weightFlag;// flag to introduce weights in IsMuon/IsMuonLoose
-  float m_dllFlag;  // flag to discriminate among the different DLLs
+  int m_dllFlag;  // flag to discriminate among the different DLLs
   
   /// GL&SF: Calculate weights:
   void P_weights(const double& p, bool *w);
